% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/cytoscape_load.R
\name{cytoscape_load}
\alias{cytoscape_load}
\title{Loads a metabolic model into cytoscape, and sets up visual mappings}
\usage{
cytoscape_load(model)
}
\description{
Loads a metabolic model into cytoscape, and sets up visual mappings
}

