% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/gene2rxn.R
\name{gene2rxn}
\alias{gene2rxn}
\alias{gene2rxn.data.frame}
\alias{gene2rxn.numeric}
\title{gene to reaction lookup}
\usage{
gene2rxn(genes, model, env = new.env(emptyenv()))

\method{gene2rxn}{data.frame}(genes, model, env = new.env(emptyenv()))

\method{gene2rxn}{numeric}(genes, model, env = new.env(emptyenv()),
  exprlist = NULL)
}
\arguments{
\item{genes}{a data frame or vector of gene activity. If a data frame, columns should be genes.}

\item{model}{a sybil model.}

\item{env}{an environment in which to evaluate the gene-reaction mappings.}

\item{exprlist}{\code{model@gprRules}, parsed to expressions. Mainly for internal use.}
}
\value{
a vector of reaction presence or activity
}
\description{
Takes a sybil model and a data frame or vector of gene activity, and
finds active reactions from this.
}
\details{
Works by creating a new environment and evaluating gpr rules
computationally. Could be a security concern, so check what you're choosing
to run. Is also quite slow, if general.
}

