% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getSelectedGenes.R
\name{get_selected_genes}
\alias{get_selected_genes}
\title{Retrieve genes from Cytoscape}
\usage{
get_selected_genes(cw, model)
}
\arguments{
\item{cw}{a cytoscape window}

\item{model}{a sybil model}
}
\value{
a character vector of gene ids
}
\description{
Retrieve the gene ids associated with nodes selected in cytoscape
}
\seealso{
Other GetSelected: \code{\link{get_selected_rxns}}
}

