% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/getCAIs.R
\name{getCAIs}
\alias{getCAIs}
\title{Calculate the CAI for each gene}
\usage{
getCAIs(DNA, anngenes,
  referenceGenes = anngenes@elementMetadata@listData$locus)
}
\arguments{
\item{DNA}{a DNAString object, giving the whole genome}

\item{anngenes}{a GRanges object, detailing how genes match to the genome}

\item{referenceGenes}{the genes selected as being highly expressed enough to form a reference set}
}
\description{
Calculate the CAI for each gene
}

