% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/heatmapify.R
\name{heatmapify}
\alias{heatmapify}
\title{Produce a heatmap}
\usage{
heatmapify(dataset, colour = "purple")
}
\arguments{
\item{dataset}{a dataframe with some columns prefixed by \code{genotype.},
and some by \code{phenotype.}}

\item{colour}{the colour for the resulting heatmap}
}
\description{
Produces a heatmap of the genes (hierachically clustered) against the first principal component of the phenotype
}

