% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/load_matlab.R
\name{load_pop_matlab}
\alias{load_pop_matlab}
\title{Load a matlab population file}
\usage{
load_pop_matlab(files, genotype_names, phenotype_names)
}
\arguments{
\item{files}{a list of files, with each row representing a member of the population, and columns representing, in order:
\enumerate{
 \item genes in the population (either as a boolean presence, or an expression value)
 \item the phenotypes
 \item rank
 \item crowding distance
}}

\item{genotype_names}{a vector of names for the genotype columns}

\item{phenotype_names}{a vector of names for the phenotype columns}
}
\value{
A \code{data.frame}, with appropriate columns prefixed by \code{genotype.} or \code{.phenotype.}
}
\description{
Loads a matlab data file (or files) representing a population
}
\examples{
\dontrun{
library (RCytoscape)
library(sybil)

# Cytoscape must be initialized, and cytoscapeRPC set up
# (See RCytoscape documentation)

cy <- CytoscapeConnection ()
cytoscape_load(Ec_core)
}
}

