% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/outlyingGenes.R
\name{outlyingGenes}
\alias{outlyingGenes}
\title{Produces graphs showing genes with unusually high correlations with position in the pareto front.}
\usage{
outlyingGenes(dataset, lowerlimit = -2, upperlimit = 2, genes_subsystems)
}
\arguments{
\item{dataset}{a dataframe with some columns prefixed by \code{genotype.},
and some by \code{phenotype.}}

\item{lowerlimit}{lower bound to mark outlying genes, in standard deviations}

\item{upperlimit}{upper bound to mark outlying genes, in standard deviations}

\item{genes_subsystems}{a data.frame with columns \code{name} and \code{SubSystem}. If this is supplied, a further graph will be produced}
}
\value{
names of outlying genes
}
\description{
Produces graphs showing genes with unusually high correlations with position in the pareto front.
}

