% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/interactive_scatterplot.R
\name{point_selector}
\alias{point_selector}
\title{Select a point from a scatterplot}
\usage{
point_selector(dataset, x, y)
}
\arguments{
\item{dataset}{a data frame}

\item{x,y}{column names from the data frame to be plotted}
}
\description{
Creates a scatter plot, and allows the user to select a point
  from it. Returns the index of the point in the dataset. See examples for typical usage
}
\examples{
point_selector(mtcars, 'hp', 'mpg')
}

