% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/interactive_scatterplot.R
\name{show_gene_expression}
\alias{show_gene_expression}
\title{Plot gene expressions in cytoscape}
\usage{
show_gene_expression(model, expressions, cw, attribute = "activity")
}
\arguments{
\item{model}{a sybil model with valid \code{react_id}s}

\item{expressions}{a vector of gene expression values, length equal to \code{length(model@allGenes)}}

\item{cw}{a cytoscape window}

\item{attribute}{the cytoscape attribute to be set}
}
\value{
silently returns the activities of the reactions as a named vector, but mainly used for the side effect of setting the 'activity' attribute in cytoscape.
}
\description{
Given a gene expressions, plot the associated reaction activities to cytoscape.
This function is a wrapper around \code{show_reaction_activities} and \code{gene2rxn}
}

