% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/interactive_scatterplot.R
\name{show_reaction_activity}
\alias{show_reaction_activity}
\title{plot reaction activities in cytoscape}
\usage{
show_reaction_activity(model, activities, cw, attribute = "activity")
}
\arguments{
\item{model}{a sybil model with valid \code{react_id}s}

\item{activities}{a named vector of reaction activities, with names from \code{model@react_ids}}

\item{cw}{a cytoscape window}

\item{attribute}{the cytoscape attribute to be set}
}
\value{
silently returns the activities of the reactions as a named vector, but mainly used for the side effect of setting the 'activity' attribute in cytoscape.
}
\description{
given a reaction activity vector, plot these to cytoscape
}

